extend VFS and IDE to support 64-bit offsets and LBA. fix MMU table memory leaks. refine exec stack ABI compliance.

+++ kernel/modules/drivers/ide.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int ide_read_sector(uint32_t lba, uint8_t *buffer);
-int ide_write_sector(uint32_t lba, uint8_t *buffer);
+int ide_read_sector(uint64_t lba, uint8_t *buffer);
+int ide_write_sector(uint64_t lba, uint8_t *buffer);
 
 #endif
